refactor: move actor to its own class
